+++ spec/paper_trail/config_spec.rb
@@ -13,40 +13,5 @@ module PaperTrail
         expect { described_class.new }.to raise_error(NoMethodError)
       end
     end
-
-    describe "#enabled" do
-      context "when paper_trail_enabled is true" do
-        it "returns true" do
-          store = double
-          allow(store).to receive(:fetch).
-            with(:paper_trail_enabled, true).
-            and_return(true)
-          allow(PaperTrail).to receive(:paper_trail_store).and_return(store)
-          expect(described_class.instance.enabled).to eq(true)
-        end
-      end
-
-      context "when paper_trail_enabled is false" do
-        it "returns false" do
-          store = double
-          allow(store).to receive(:fetch).
-            with(:paper_trail_enabled, true).
-            and_return(false)
-          allow(PaperTrail).to receive(:paper_trail_store).and_return(store)
-          expect(described_class.instance.enabled).to eq(false)
-        end
-      end
-
-      context "when paper_trail_enabled is nil" do
-        it "returns true" do
-          store = double
-          allow(store).to receive(:fetch).
-            with(:paper_trail_enabled, true).
-            and_return(nil)
-          allow(PaperTrail).to receive(:paper_trail_store).and_return(store)
-          expect(described_class.instance.enabled).to eq(true)
-        end
-      end
-    end
   end
 end