+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Breaking Changes
 
+- `PaperTrail.enabled=` now affects all threads
 - [#556](https://github.com/airblade/paper_trail/pull/556) /
   [#301](https://github.com/airblade/paper_trail/issues/301) -
   If you are tracking who is responsible for changes with `whodunnit`, be aware